Adjusted silkcreen and linked the STEP files.
The silkscreen was adjusted to reflect the correct dimensions of the display. 

The linked step files are adjusted for location and scale.

  (model /home/user/kicad/libraries/User_3Dshapes/sharp.STEP
    (offset (xyz 25.65 15.3 -124.5))
    (scale (xyz 1 1 2))
    (rotate (xyz 0 0 0))
  )
  (model /home/user/kicad/libraries/User_3Dshapes/connecctor.STEP
    (offset (xyz 27.15 4.8 -2.75))
    (scale (xyz 1 1 1))
    (rotate (xyz -90 0 0))
  )

In order for the STEP files to show on your footprint, you have to change the path "/home/user/kicad/libraries/User_3Dshapes/" to the path you have the STEP files.